refaktorisering av medlemskapsperioder
